# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SchoolPilot API
  description: |
    Multi-tenant school management SaaS API.
    Authenticated routes require `Authorization: Bearer <accessToken>`.
    OpenAPI spec also served at `GET /api/docs/openapi.yaml`.
  version: 1.0.0
  contact:
    name: Great Goodluck Schools
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.yourschoolpilot.com
    description: Production
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Returns accessToken and user
        '401':
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token (cookie or body)
      responses:
        '200':
          description: New accessToken
        '401':
          description: Invalid refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT